# Improve performance by increasing the buffer size to 20K