@arp242 2023-10-01
Can also start at 1K like before, and then grow to 20K after we've done a few loops (and after that even further): that way memory usage stays small for small files, but is still fast for larger files.

I don't really know what your expectations for this are.