@cktan 2023-10-08
In my view, TOML is primary meant for configuration. Most TOML files probably have less than 100 lines. 1K seems very small, and I don't have objections to grow it to around 16k.