@arp242 2023-10-08
Yes, I agree. It's just that it would be nice to also make it work reasonably well with larger files, and seemed a simple enough change.

I changed it to 16K now.